remove unnecessary cast

+++ bundles/org.eclipse.swt/Eclipse SWT PI/carbon/org/eclipse/swt/internal/carbon/OS.java
@@ -765,8 +765,8 @@ public class OS extends C {
 	public static final int kPMDestinationFile = 2;
 	public static final int kPMDestinationPreview = 4;
 	public static final int kPMDestinationPrinter = 1;
-	public static final int kPMLandscape = 2;
-	public static final int kPMPortrait = 1;
+	public static final short kPMLandscape = 2;
+	public static final short kPMPortrait = 1;
 	public static final int kPMPrintAllPages = 2147483647;
 	public static final int kPMShowDefaultInlineItems = 1 << 15;
 	public static final int kPMShowPageAttributesPDE = 1 << 8;
